Early version of die vision module
The image that I tested this on has also been pushed as well. No further
images will be pushed in an effort to save space on the repository.

Can recognize and distinguish between multiple die in the same image and
identify the pips on them. Struggles when the pips are malformed. Need
to see how this will work underwater.