Fix the SHM bug!
Problem #1: robust mutices were disabled because there was a glibc
version check - there was a critical bug that could cause deadlocks in
glibc 2.25, but it has since been fixed.

Problem #2: watcher condition variables were not being correctly cleaned
up on process death. Something must have changed in the undocumented
semantics of pthread condition variables. But basically SHM watchers
could hang in pthread_cond_broadcast because the internal mutex in the
watcher's condition variable was still held by the dead process. It is
important to note that this condition variable is tied to the process; the
problem is that we recycle memory for watchers, so subsequent processes
that were given the same block experienced the problem. Fixed by
re-initializing the condition variables every time a watcher is created
(which should have been done from the beginning).